pyproject.toml
[project]
name = "ai-news-agent"
version = "0.1.0"
description = "AI news agent that crawls sources, synthesizes digests with Claude, and emails them"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "anthropic>=0.40",
    "resend>=2.0",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.scripts]
news = "ai_news_agent.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_news_agent"]
